# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Samad Specs API",
    "version": "1.0.0",
    "description": ""
  },
...
  "components": {
...
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "username": {
            "type": "string",
            "example": "johnll21"
          },
          "password": {
            "minLength": 8,
            "type": "string",
            "example": "somepassword8"
          },
          "role": {
            "type": "string",
            "example": "user | admin"
          }
        }
      },
...
    },
    "securitySchemes": {
      "accessToken": {
        "in": "cookie",
        "type": "apiKey",
        "name": "Access Token",
        "description": "JWT access token stored in a cookie named `id`"
      },
      "refreshTokenAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "Refresh Token",
        "description": "JWT access token stored in a cookie named `rid`"
      }
    }
  },
  "paths": {
...
    "/api/v1/user/{id}": {
      "put": {
        "tags": ["User"],
        "summary": "Update a user",
        "security": [
          {
            "accessToken": []
          },
          {
            "refreshTokenAuth": []
          }
        ],
        "x-roles": ["Admin"],
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Ok, dont have response"
          },
          "400": {
            "description": "Bad Request, validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "code": 400,
                    "message": "Bad Request",
                    "errors": "Comma separated Validation error"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "code": 400,
                    "message": "Not found",
                    "errors": "Update failed"
                  }
                }
              }
            }
          },
          "500": {
            "description": "internal server error"
          }
        }
      },
      "delete": {
        "tags": ["User"],
        "summary": "Update a user",
        "security": [
          {
            "accessToken": []
          },
          {
            "refreshTokenAuth": []
          }
        ],
        "x-roles": ["Admin"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted, dont have response"
          },
          "404": {
            "description": "Not found resource to update",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "code": 404,
                    "message": "Bad Request",
                    "errors": "Delete failed"
                  }
                }
              }
            }
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
...
  }
}